+++ ksolve/CMakeLists.txt
@@ -26,5 +26,6 @@ add_library(ksolve
         SteadyState.cpp
         Gsolve.cpp
         ZombiePoolInterface.cpp
+        BoostSystem.cpp
         testKsolve.cpp
     )
