+++ ksolve/OdeSystem.h
@@ -10,54 +10,7 @@
 #ifndef _ODE_SYSTEM_H
 #define _ODE_SYSTEM_H
 
-#ifdef USE_BOOST
-
-#include <boost/numeric/odeint.hpp>
-typedef boost::numeric::ublas::vector< double > vector_type_;
-typedef boost::numeric::ublas::matrix< double > matrix_type_;
-
-/*
- * =====================================================================================
- *        Class:  BoostSys
- *  Description:  The ode system describing chemical kinetics in BOOST.
- * =====================================================================================
- */
-class BoostSys
-{
-    public:
-        BoostSys (){ };                             /* constructor */
-
-        /*-----------------------------------------------------------------------------
-         *  Following functiors implement equivalent of 'function' and
-         *  `jacobian` of gsl_odeiv2_system. These wrappers are just to have
-         *  consistency between calls to gsl or boost solver.
-         *-----------------------------------------------------------------------------*/
-        int (*rhs) (  double y
-                , double& dydt
-                ,  double t 
-                //, void * params 
-                );
-
-        // Fixme: Change the types of argument.
-        int (*jacobian) ( double t
-                , const double y[]
-                , double dfdt[]
-                , void* params
-                );
-
-        size_t dimensions;                      /* dimensions of the system */
-
-        /* Pointer to the arbitrary parameters of the system */
-        void * params;
-
-        boost::numeric::odeint::runge_kutta_dopri5< double, double ,double
-            , double, boost::numeric::odeint::vector_space_algebra > stepper;
-
-}; /* -----  end of class BoostSys  ----- */
-
-
-#endif
-
+#include "BoostSystem.h"
 
 class OdeSystem {
 	public:
